Add WaniKani info panel width stylesheet
This is a stylesheet I wrote to limit the width of the info panel on WaniKani
lessons and reviews to 50rem, making explanations a little easier to read

By default, the info panel on lessons and reviews spans almost the full width of
your window. For windows larger than around 800px, this results in a sub-optimal
reading experience, particularly for meaning and reading explanations. This is
both a usability and accessibility issue

This is slightly modified from the version uploaded to Stylish/userstyles.org at
the link below, taking out the Stylish variable interpolation. It does still
include the deprecated `@-moz-document` selector, which is used by Stylish to
scope the styles just to the given URLs

There's also an automatically generated JavaScript implementation at the link
below - this is provided by Stylish so don't blame the old-school syntax on me!

@see https://userstyles.org/styles/257533/wanikani-lesson-review-info-panel-width